Add authorization

+++ src/offer/offer.controller.js
@@ -1,5 +1,6 @@
 import { Offer } from './offer.model.js';
 import flatten from 'flat';
+import { USER_ROLE } from '../constants.js';
 
 export const getMany = async (req, res) => {
   const offers = await Offer.find({});
@@ -8,7 +9,9 @@ export const getMany = async (req, res) => {
 
 export const getOne = async (req, res) => {
   if (!req.params.id)
-    res.status(400).json({ message: 'Bad request: no ID param', errors: [] });
+    return res
+      .status(400)
+      .json({ message: 'Bad request: no ID param', errors: [] });
   try {
     const offer = await Offer.findById(req.params.id);
     res.status(200).json({ data: offer });
@@ -22,7 +25,15 @@ export const getOne = async (req, res) => {
 
 export const updateOne = async (req, res) => {
   if (!req.params.id)
-    res.status(400).json({ message: 'Bad request: no ID param', errors: [] });
+    return res
+      .status(400)
+      .json({ message: 'Bad request: no ID param', errors: [] });
+
+  const offer = await Offer.findById(req.params.id);
+
+  if (!offer.host.equals(req.user._id) && req.user.role !== USER_ROLE.ADMIN)
+    return res.status(400).json({ message: 'Access denied', errors: [] });
+
   try {
     const flattenedBody = flatten(req.body);
     const updatedOffer = await Offer.findByIdAndUpdate(
@@ -41,7 +52,15 @@ export const updateOne = async (req, res) => {
 
 export const removeOne = async (req, res) => {
   if (!req.params.id)
-    res.status(400).json({ message: 'Bad request: no ID param', errors: [] });
+    return res
+      .status(400)
+      .json({ message: 'Bad request: no ID param', errors: [] });
+
+  const offer = await Offer.findById(req.params.id);
+
+  if (!offer.host.equals(req.user._id) && req.user.role !== USER_ROLE.ADMIN)
+    return res.status(400).json({ message: 'Access denied', errors: [] });
+
   try {
     const removedOffer = await Offer.findByIdAndRemove(req.params.id);
     res.status(200).json({ data: removedOffer });
@@ -55,7 +74,10 @@ export const removeOne = async (req, res) => {
 
 export const createOne = async (req, res) => {
   try {
-    const createdOffer = await Offer.create(req.body);
+    const createdOffer = await Offer.create({
+      ...req.body,
+      host: req.user._id,
+    });
     res.status(200).json({ data: createdOffer });
   } catch (error) {
     res
